Add files via upload
Day 2 Afternoon Assignment (going back and forth between Python and Unix) 

Unix-commands.sh contain commands that I ran from my terminal in my ~/qb25-answers/unix-python-scripts folder with copies of all necessary data files in the working directory. These commands answer the questions in parts 1, 3, and 5.

recalculate-score.py was ran from terminal in the same folder and generated new_scores_ce11_genes.bed, and it is the script for question 2

sample-by-gene.py is the script for question 4

gtf2bed.py is the script for question 6 and it generated gencode.v46.basic.annotation.bed (from a gzipped file with the same name and a .gtf extension).